third_party/blink/web_tests/external/wpt/css/cssom-view/scrollIntoView-direction-rtl-expected.txt

This is a testharness.js-based test.
FAIL scrollIntoView({inline: "start"}), direction: rtl assert_approx_equals: start should be the right edge expected -200 +/- 0.5 but got 200
FAIL scrollIntoView({inline: "center"}), direction: rtl assert_approx_equals: should center the target expected -150 +/- 0.5 but got 150
FAIL scrollIntoView({inline: "end"}), direction: rtl assert_approx_equals: end should be the left edge expected -100 +/- 0.5 but got 100
Harness: the test ran to completion.
